Fix ending lba problem
When an chromium os image is smaller than the storage device (say, SD card),
the value of dev_desc->lba is not the LBA of backup GPT header, but the last
block of the storage device.

Therefore we have to look up the LBA of backup GPT header in primary GPT
header (LBA = 1).

BUG=chromium-os:12124
TEST=manual

1. Create a chromiumos image that is smaller than SD card capacity
2. Run u-boot scripts: "cros load_k ${loadaddr} 0x01000000 0x10000000"
3. Then check debug message to see whether it correctly set ending_lba

Review URL: http://codereview.chromium.org/6686041

Change-Id: I215258d034e5a46bcda5f0a0064ddc670b097d16