No retry for USB storage to speed up the resume time.
In the scenario, a user inserts a USB storage and boots to recovery mode, we
required the user to plug the USB out first. However, in the meantime that the
system is scanning the USB, the user plugs it out. It may need ~50sec to resume
because there are many retry there. This change makes it no retry such that it
only takes <1sec to resume.

R=clchiou@chromium.org,sjg@chromium.org
BUG=chromium-os:10500
TEST=emerge-tegra2_seaboard chromeos-u-boot-next successfully.
When the system is scanning the USB storage, plug the USB out.
It takes <1sec to resume.

Review URL: http://codereview.chromium.org/6685083

Change-Id: I1ed1d54c211dda87693b91302586ae1f534d75aa