Fix Seaboard UART corruption on SPI activity
On Seaboard the UART and SPI interfere with each other. This causes the UART
to receive spurious zero bytes after SPI transactions and also means that
SPI can corrupt a few output characters when it starts up if they are still
in the UART buffer.

This hack corrects this by making SPI record that it may have corrupted the
UART, and making the UART take evasive action.

BUG=chromium-os:13228
TEST=Try developer U-Boot on Seaboard, make sure it auto-boots OK now

Change-Id: I130f5e8bdf1be48b20015c8b2228d972409cc135

Review URL: http://codereview.chromium.org/6715017