Store VbNvContext in MBR
The VbNvContext cookie has to be stored in a place that firmware and
kernel can access even before the partition table is created. So the EFI
partition is not a good candidate of storage.

On ARM platforms the MBR is not used by any part of the system, and so
it is quite safe to use MBR for VbNvContext storage.

This commit also reverts commit 8e23ac.

Note: The VbNvContext was stored in SPI flash, but it turned out that
kernel could not access SPI flash due to driver issues. So the storage
was temporarily moved to eMMC device.

R=rongchang@chromium.org
BUG=none
TEST=none

Review URL: http://codereview.chromium.org/6878095

Change-Id: Ie460318a653c54060dfc238a8d3bd77ade16c59b