Allow to create shared memory space without loading kernel.
This change provides the ability to create the shared
memory blob without invoking the verified boot Loadkernel()
function.

Setting environment variable `bypass_load_kernel' to any
value will prevent invoking Loadkernel() by the
'cros load_k' cli command.

This makes debugging the shared memory issues much easier
as the user can load the kernel by other means (like
netbooting) but still have the shared memory structure in
place.

This CL also rearranges the shared memory structure (such
that fields the kernel driver cares about are moved to the
base of the structure) and introduces a field to convey the
size of the shared memory structure. This allows the kernel
driver to be oblivious to the actual data layout of the
remainder of the structure and to make potential future
layout changes  transparent for the driver.

Change-Id: I03d879c180ad5869bab6c2e21d92c5a8d90c0c43

BUG=None
TEST=manual, see below:

The test requires availability of the kernel on a tftp host,
proper kernel command line in the environment and root file
system present on a drive.

. program the updated u-boot image on kaen
. restart it and catch in u-boot shell
. enter the following command sequence:
  setenv bypass_load_kernel 1
  usb start
  dhcp
  tftpboot ${loadaddr} server:path
  cros load_k 1 2  # using dummy parameters 1 and 2
  bootm
. wait for the target to start up
. install the chromeos_arm module and observe it report
  finding the signature.

Review URL: http://codereview.chromium.org/6902086