CHROMIUM: unroll the outer for-loop of VbExDiskGetInfo()
VbExDiskGetInfo() was implemented with nested loops, where the outer
loop iterates iterators, and the inner loop runs through block devices
via the iterator.

It was quite complicated and the outer loop only iterates twice. So we
believe it will make the code more readable if we unroll the outer
for-loop.

Besides unrolling the loop, we fix a logic error that disk_flags should
be tested 'is subset of' rather than 'has non-empty intersection with'
flags.

BUG=chromium-os:16542
TEST=vbexport_test diskinfo

Change-Id: Ic4a7d516d91ba7214b9b67eebbecf02f80b11786
Signed-off-by: Tammo Spalink <tammo@chromium.org>
Reviewed-on: http://gerrit.chromium.org/gerrit/4472
Tested-by: Che-Liang Chiou <clchiou@chromium.org>
Reviewed-by: Tom Wai-Hong Tam <waihong@chromium.org>