Add a default x86 implementation for cleanup_before_linux.
This function provides an opportunity for some last minute cleanup and
reconfiguration before control is handed over to Linux. It's possible this
may need to do something in the future, but for now it's left empty. It's set
up as a weak symbol so it can be overridden if necessary on a case by case
basis.

BUG=chrome-os-partner:4552
TEST=Built and booted u-boot on an Alex.

Change-Id: Ia59436b2a9d499ff659b2f41668cf951f2f7d999
Signed-off-by: Gabe Black <gabeblack@google.com>
Reviewed-on: http://gerrit.chromium.org/gerrit/4698
Reviewed-by: Stefan Reinauer <reinauer@google.com>
Tested-by: Gabe Black <gabeblack@chromium.org>