mmc: tegra2: fix HC version register offset
The ADMA address register has provision to support hardware with 64-bit
physical addressing. By the way, on Tegra2 a long is 32 bits, so we need
to reserve another 32-bit word to ensure that the next registers are at
the right place.

BUG=none
TEST=run U-Boot on Qemu and check that the host controller version is
read at offset 0xfe and no longer at 0xfa

Change-Id: I8de2e1e25f29cfe7ba59998c5cb1d711661b280e
Reviewed-on: http://gerrit.chromium.org/gerrit/4469
Reviewed-by: Tom Warren <twarren@nvidia.com>
Tested-by: Vincent Palatin <vpalatin@chromium.org>
Reviewed-by: Simon Glass <sjg@chromium.org>