Conditionally implement vboot function wipe_unused_memory
The wipe_unused_memory function in bootstub_entry.c in the vboot library which
depends on being able to call fdt_decode_memory which only compiled in with
CONFIG_OF_CONTROL. That setting breaks booting u-boot with coreboot right now.

This change conditionally compiles in the existing implementation if the
right support is available and compiles in a warning message if it isn't.

BUG=chrome-os-partner:4552
TEST=Built on x86.

Change-Id: I6404eb12b25c19eb21eb65ad16a5031a5e3d0ad7
Reviewed-on: http://gerrit.chromium.org/gerrit/4843
Reviewed-by: Stefan Reinauer <reinauer@google.com>
Tested-by: Gabe Black <gabeblack@chromium.org>