CHROMIUM: remove CONFIG_HARDWARE_TPM
The CONFIG_HARDWARE_TPM flag was used to disable software emulation of
TPM functionality, for testing purpose or for boards that have a buggy
or even do not have a hardware TPM. However, it turns out that it is
easier to mock out TPM entirely in vboot_reference than implement a full
software-emulated TPM in U-Boot. So we believe remove this flag and
implement a mock in vboot_reference.

BUG=none
TEST=boot on Aebl

Change-Id: Id6d2975f04f0f5e4203d45e23f36fce2ad7114f3
Reviewed-on: http://gerrit.chromium.org/gerrit/4647
Tested-by: Che-Liang Chiou <clchiou@chromium.org>
Reviewed-by: Rong Chang <rongchang@chromium.org>