Initialize the device tree pointer to NULL in i386.
If the device tree pointer isn't set to anything, which could happen if the
device tree isn't found, for instance, or simply never set up, the pointer
could be set to something arbitrary. This change forces initializes it to NULL
so that it's more obviously not set.

BUG=chrome-os-partner:4993
TEST=Built and booted u-boot on coreboot, used temporary printfs to print the
device tree pointer when booting chromeos, and verified that the pointer was
NULL if not set.

Change-Id: I9c6c34da6f3309b31a1fda920c6b139d728561f3
Signed-off-by: Gabe Black <gabeblack@google.com>
Reviewed-on: http://gerrit.chromium.org/gerrit/4884
Reviewed-by: Stefan Reinauer <reinauer@google.com>
Tested-by: Gabe Black <gabeblack@chromium.org>