Tell the FDT library where the device tree is automatically on bringup
This change adds a call to set_working_fdt_addr near the end of u-boot
initialization which tells the fdt command/library where the device tree is.
This makes it possible to use the fdt command to look at the active device tree
since otherwise there would be no way to know what address it was at to set
things up manually.

BUG=chrome-os-partner:4993
TEST=Built for coreboot/Alex, booted, and used fdt to look at the device tree.

Change-Id: Iccd93af0c3934de5f284b76f8c894adc9dbd7f88
Signed-off-by: Gabe Black <gabeblack@google.com>
Reviewed-on: http://gerrit.chromium.org/gerrit/4919
Reviewed-by: Gabe Black <gabeblack@chromium.org>
Tested-by: Gabe Black <gabeblack@chromium.org>