Increase the size of the heap to 4MB
This change increases the size of the heap on i386/coreboot to 4MB to match
the Tegra configurations. The extra size is needed when allocating space for
copies of the read/write firmware in memory so that signatures can be checked,
etc.

BUG=chrome-os-partner:4552
TEST=Built and booted on x86-alex. Ran vboot_twostop and saw previously
failing allocations succeed. Booted non-verified ChromeOS to the login screen.

Change-Id: I075604975c90455eb21527d8153ca05a39dafa4c
Signed-off-by: Gabe Black <gabeblack@google.com>
Reviewed-on: http://gerrit.chromium.org/gerrit/5386
Reviewed-by: Gabe Black <gabeblack@chromium.org>
Tested-by: Gabe Black <gabeblack@chromium.org>