Fix checking if information about vboot's nv storage was in the cb tables
u-boot was not checking for information about vboot's non-volatile storage
correctly. If the information taken from the coreboot table was non-zero (when
there was something there), u-boot would fail, which is essentially the
opposite of correct. This change sets sentry values in the structure filled
from coreboot's tables and then checks for those to determine if there's any
information about nonvolatile storage.

BUG=chrome-os-partner:4552
TEST=Built and ran vboot_twostop. Verified that attempts to access nonvolatile
storage no longer fail.

Signed-off-by: Gabe Black <gabeblack@google.com>

Change-Id: I254c5808637dcd668c0529bbcfa3ba22924165b7
Reviewed-on: http://gerrit.chromium.org/gerrit/5542
Reviewed-by: Gabe Black <gabeblack@chromium.org>
Tested-by: Gabe Black <gabeblack@chromium.org>