Wrap small helper functions from libgcc to avoid an ABI mismatch
When gcc compiles some 64 bit operations on a 32 bit machine, it generates
calls to small functions instead of instructions which do the job directly.
Those functions are defined in libgcc and transparently provide whatever
functionality was necessary. Unfortunately, u-boot and vboot are both built
with a non-standard ABI and libgcc isn't. When the two are linked together,
very confusing bugs can crop up, for instance seemingly normal integer
division or modulus getting the wrong answer or even raising a spurious
divide by zero exception.

This change barrows (steals) a technique and some code from coreboot which
solves this problem by creating wrappers which translate the calling
convention when calling the functions in libgcc. Unfortunately that means that
these instructions which had already been turned into functions have even more
overhead, but more importantly it makes them work properly.

To find all of the functions that needed wrapping, u-boot was compiled without
linking in libgcc. All the symbols the linker complained were undefined were
presumed to be the symbols that are needed from libgcc. These were a subset of
the symbols covered by the coreboot code, so it was used unmodified.

It wouldn't be a bad idea to revisit vboot and try to either get rid of these
slow operations, or make them faster by enabling appropriate instructions in
gcc. Currently the vboot Makefile sets -march=i386 which is a very conservative
setting. Something more aggressive might help.

BUG=chrome-os-partner:4552
TEST=Built and ran vboot_twostop. Saw previous incorrect behavior go away.

Signed-off-by: Gabe Black <gabeblack@google.com>

Change-Id: I45439fd8e967754b9168700f152fbab09521c99f
Reviewed-on: http://gerrit.chromium.org/gerrit/5543
Reviewed-by: Gabe Black <gabeblack@chromium.org>
Tested-by: Gabe Black <gabeblack@chromium.org>