Refactor the zboot underpinnings so they can be reused with a vboot image
If vboot successfully verifies a kernel, it will leave it in place and
basically ready to boot. The zeropage table which is part of the x86 boot
protocol is at the end of the kernel, though, instead of the beginning, and
because the image is already in place there's no need to copy it around. This
change refactors the code which implements the zboot command so that the
configuration of the zeropage table and loading the pieces of the kernel into
memory are done separately. Also, because the command line goes before the
zeropage table in vboot which is somewhat incompatible with the normal
protocol, where to put the command line is a now a parameter instead of being
hard coded.

BUG=chrome-os-partner:4552
TEST=Built and booted a legacy kernel with both the 32 bit and 16 bit boot
protocols. Like other, earlier changes to this code only a current kernel which
doesn't exercise old protocol features was tested. This change is more
significant than the others and so may break booting old kernels. Also, the
bootm command was not tested on x86. We need to decide if that command makes
any sense to keep/support.

Signed-off-by: Gabe Black <gabeblack@google.com>

Change-Id: I09f6d0e2d46b4ce499f31c8073571968dba51b8a
Reviewed-on: http://gerrit.chromium.org/gerrit/5545
Reviewed-by: Simon Glass <sjg@chromium.org>
Tested-by: Gabe Black <gabeblack@chromium.org>