Modify the API for doing transfers over SPI.
This change modifies the spi_xfer function to take read and write buffers and
read and write bit counts which is similar to what flashrom does. This has the
nice effect of simiplifying many call sights, and will allow integrating SPI
controllers which don't allow arbitrary control of the chip select lines.

Only boards which are being used with ChromeOS and the functionality they use
will be updated immediately. That's because those are the cases which can
actually be tested effectively, and fixing up the other locations can be put
off until later. These change and the eventual change that updates all the
other call sights/drivers can also be reviewed and tested independently. It
will also make rebasing easier.

To try to make the transition less painful, the new interface is selectable
using the CONFIG_NEW_SPI_XFER option. This selects whether the new or old
interface is used.

BUG=chrome-os-partner:4722
TEST=With this and related changes, built for x86-alex, tegra2_kaen,
tegra2_aebl, tegra2_seaboard, and tegra2_asymptote. Flashed the firmware on a
Kaen, verified the checksums matched, and booted to ChromeOS login. Built with
the option turned off on x86-alex and tegra2_kaen.

Change-Id: If32843255584544b08b9c8e18bbd9f36273fb71b
Signed-off-by: Gabe Black <gabeblack@google.com>
Reviewed-on: http://gerrit.chromium.org/gerrit/5374
Reviewed-by: Stefan Reinauer <reinauer@google.com>