Add a new Tegra SPI driver which uses the new spi_xfer interface
This interface is different enough that it warrants a new driver, not a
retrofit of the old one. Fortunately, flashrom had already taken the Tegra SPI
driver from u-boot and modified it to use essentially this new interface. This
change ports it back again and makes the interface selection setting pick which
driver to use.

BUG=chrome-os-partner:4722
TEST=With this and related changes, built for x86-alex, tegra2_kaen,
tegra2_aebl, tegra2_seaboard, and tegra2_asymptote. Flashed the firmware on a
Kaen, verified the checksums matched, and booted to ChromeOS login. Built with
the option turned off on x86-alex and tegra2_kaen.

Change-Id: Ia2ce772a8f8f3194ff8a76e1abe0e07061fa5672
Signed-off-by: Gabe Black <gabeblack@google.com>
Reviewed-on: http://gerrit.chromium.org/gerrit/5618
Reviewed-by: Simon Glass <sjg@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Stefan Reinauer <reinauer@google.com>