fdt: expose GPIO decoder to public interface
fdt_decode module is able to decode GPIO configurations (GPIO number,
I/O direction, and polarity). But the GPIO decoder was kept private to
fdt_decode module because GPIOs were generally used to control a larger
device (LCD, for example) and so considering a single GPIO in isolation
usually did not make much sense.

However, we are aware that there are cases that a GPIO is not a part of
any other larger device, but a standalone prober of hardware status. In
these cases, decoding a single GPIO does make sense. Thus, we expose the
GPIO decoder to public interface.

As this patch only changes functions from static to non-static and
function names, I only tested it builds or not.

BUG=none
TEST=build for Aebl successfully

Change-Id: Ic5f19b64f0cf3f64c56adead7c57e480ef24cab8
Reviewed-on: http://gerrit.chromium.org/gerrit/5725
Tested-by: Che-Liang Chiou <clchiou@chromium.org>
Reviewed-by: Anton Staaf <robotboy@chromium.org>
Reviewed-by: Simon Glass <sjg@chromium.org>