Fix compiler warnings generated by gcc 4.6 (local only).
The newer compiler version is more thorough in detecting code
inconsistencies, and reports warnings in many cases when building
u-boot for kaen and alex.

This change modifies the not yet upstreamed files to get rid of
the warnings. There supposed to be no logical changes, so no
testing other than building the system is being done.

BUG=chromium-os:18862
TEST=manual
  . run the following commands

    emerge-tegra2_kaen chromeos-u-boot
    emerge-x86-alex chromeos-u-boot

  . observe them succeed

Change-Id: I7b8f3e7211007537f9e85cfd059174dbb9763bb4
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>
Reviewed-on: http://gerrit.chromium.org/gerrit/5771
Reviewed-by: Simon Glass <sjg@chromium.org>