Add 'sf update' command to do smart SPI flash update
This adds a new SPI flash command which only rewrites blocks if the contents
need to change. This can speed up SPI flash programming when much of the
data is unchanged from what is already there.

To implement this, a new probe_block method is added to the flash API. This
returns information on the erase block size at a given offset.

BUG=chromium-os:17805
TEST=build and flash on Aebl with modified flash script. See that flashing
is performed more quickly. Time before = 72 seconds; time after 26 seconds

Change-Id: I3382553c9a06fabcddc00f1973b3b5e4e2d2f72d
Reviewed-on: http://gerrit.chromium.org/gerrit/6157
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Tested-by: Simon Glass <sjg@chromium.org>