elog: Pass event log info to the kernel, and trim out the panic buffer
Pass the location of the event log panic buffer and the SPI flash
offset of the event log itself to the kernel through its device tree.

Also, trim the panic buffer range out of the /memory/reg property to
keep the kernel from using it for something else.

This is a cherry pick of b467622 from the 4482 branch, there was an
easily resolved conflict in cros/lib/boot_kernel.c:ft_board_setup().

BRANCH=none
BUG=chrome-os-partner:16132
TEST=Built and booted on pit. With prototyped panic code, verified
     that a panic event was stored in the right place, and was picked
     up by the firmware and recorded. Looked at the device tree in the
     booted system and verified that the expected properties where
     present in the right places.

Change-Id: Ia2041c9a836540316534f5ca582825c94844b5d8
Signed-off-by: Gabe Black <gabeblack@google.com>
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/66782
Reviewed-by: Simon Glass <sjg@chromium.org>
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/186473
Reviewed-by: Doug Anderson <dianders@chromium.org>