rtc: Add rtc type to the machine params
The eventlog code needs the RTC in SPL, and the RTC needs to be able to tell
if it should run.

BUG=chrome-os-partner:16132
TEST=With this and other changes, built and booted on Pit. Verified that
a wake event is recorded with a time stamp, although the rtc driver can't talk
to the RTC over I2C after a resume for some reason.
BRANCH=pit

Change-Id: I02a207eed46e3e41e4d4fd14bc19f9d11aa5273e
Signed-off-by: Gabe Black <gabeblack@google.com>
Reviewed-on: https://gerrit.chromium.org/gerrit/65170
Reviewed-by: Gabe Black <gabeblack@chromium.org>
Tested-by: Gabe Black <gabeblack@chromium.org>
Commit-Queue: Gabe Black <gabeblack@chromium.org>