efs: cros: config: Adjust configuration for the small builds
Create a cut-down CROS_RO build for use by early-firmware-selection. This
allows U-Boot to run from IRAM.

Assume a CROS_FULL build unless CROS_SMALL is defined so that the full
U-Boot is the default in cases where no CROS_... defines are given.
This preserves the expected behavior.

BUG=chrome-os-partner:21115
BRANCH=pit
TEST=manual
Build for each of CROS_FULL, CROS_SMALL, CROS_RO, CROS_RW and see that
all builds succeed and boot on link in normal mode.
Along with changes, build and boot a full RW U-Boot flow with
early-firmware-selection on Pit.

Change-Id: I76fc9bf14d91eeda3d05c999d66283e99fe3384b
Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/64102