video: Add CONFIG_VIDEO_NO_TEXT to remove text/fonts
This option permits a graphics-only console with no text. This removes the
U-Boot font from the image and reduces the code size somewhat.

(This option would be better implemented across both LCD and video, but
for now it just supports video).

BUG=tbd
TEST=manual
Build and boot normal verified boot flow on link. Build and boot into
recovery and see that the screen appears correctly.

Change-Id: Ic00b19e717aa4d6e751be44ac19e60c9855baf0b
Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/64103
Reviewed-by: Che-Liang Chiou <clchiou@chromium.org>