sandbox: Add os_jump_to_image() to run another executable
For some tests it is useful to be able to run U-Boot again but pass on the
same memory contents. Add a function to achieve this.

BUG=chromium:244019
TEST=manual
With other changes, see that chromeos_sandbox can perform a developer
mode verified boot flow from start to finish.
Signed-off-by: Simon Glass <sjg@chromium.org>

Change-Id: I6ac7d6cca78ba50c27769496cddef68f496be33b
Reviewed-on: https://gerrit.chromium.org/gerrit/64115
Commit-Queue: Simon Glass <sjg@chromium.org>
Reviewed-by: Simon Glass <sjg@chromium.org>
Tested-by: Simon Glass <sjg@chromium.org>