gen: sandbox: Add LCD driver
Add a simple LCD driver which uses SDL to display the image. We update the
image regularly, while still providing for reasonable performance.

Adjust the common lcd code to support sandbox.

For command-line runs we do not want the LCD to be displayed, so add a
--hide_lcd option to disable it.

BRANCH=none
BUG=chromium:244019
TEST=manual
./u-boot -d u-boot.dtb

See that it comes up with a display window now, and it is possible to type
at the prompt on the display.

Change-Id: I66751c58005b6c186492b07c10c130872b27a926
Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/64122
Tested-by: Che-Liang Chiou <clchiou@chromium.org>