gen: cros_ec: Support systems with no EC interrupt
Some systems do not have an EC interrupt. Rather than assuming that the
interrupt is always present, and hanging forever waiting for more input,
handle the missing interrupt. This works by reading key scans only until
we get an identical one. This means the EC keyscan FIFO is empty.

BRANCH=none
BUG=chromium:244019
TEST=manual
Run sandbox U-Boot with the cros_ec keyboard enabled in the config, and
see that it no longer hangs on boot.

Change-Id: Iad6190611ca68bbbd783de29222d0da0c550b8b7
Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/64123
Tested-by: Che-Liang Chiou <clchiou@chromium.org>