gen: sandbox: Fix verbosity of sandbox host block device
We don't want to print a message in the driver every time we find a missing
device. Pass the information back to the caller where a message can be printed
if desired.

BRANCH=none
BUG=chromium:244019
TEST=manual
With other changes, run a verified boot into recovery mode and see that the
output is no longer spamed by mising device messages.

Change-Id: Ie22119f43cdf03866f671aa489c55d55c7baa48b
Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/64128