gen: sandbox: Allow Ctrl-C to work in sandbox
It is useful for Cltl-C to be handled by U-Boot as it is on other boards.
But it is also useful to be able to terminate U-Boot with Ctrl-C.

Add an option to enable signals while in raw mode, and make this the
default. Add an option to leave the terminal cooked, which is useful for
redirecting output.

BRANCH=none
BUG=chromium:244019
TEST=manual
Try running with the various options and see that they behave as expected.

Change-Id: Ie9a7bdc8f4057c19d4d32cf7636fe92cd31e764f
Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/64129