gen: mmc: get MMC device index from FDT node
Define and implement a function to get an MMC device from an MMC
node in the device tree.  Similar to the SPI spi_get_bus_by_node(),
this provides a platform-independent way to find, and subsequently
initialize an MMC node from the device tree.

This adds an FDT node field to the MMC struct, which is set by the
platform-specific MMC driver, and is used to find the matching MMC
device.

Setup node field for the Exynos DWMMC driver.

BUG=none
BRANCH=none
TEST=With vboot MMC support, successfully boot with vboot from eMMC.
The MMC device is correctly determined by mmc_get_device_by_node().

Change-Id: I5c08da16c9eaf41afabeec821ffad2618be2c683
Signed-off-by: Michael Pratt <mpratt@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/63681
Reviewed-by: Simon Glass <sjg@chromium.org>