gen: Relocate only the used size of the FDT
This is a small bug. We allow the control FDT to grow slightly when
relocating it. But when copying the FDT from the old place we use the
new size instead of the old. Correct this.

BUG=chrome-os-partner:21115
BRANCH=none
TEST=manual
Build and boot on pit and sandbox, see that FDT is still available.

Change-Id: If4ea863bd12a01621f4b49675f81d56f4e8a3d4b
Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/64131