efs: arm: Remove CONFIG_SYS_SYM_OFFSETS
Now that ARM relocation has been adjusted to remove absolute relocations,
we don't need this feature. Remove it so that ARM works in a similar way to
other architectures.

This change is needed to avoid adding special cases for ARM in various
places in the code.

BUG=chrome-os-partner:21115
BRANCH=pit
TEST=manual
Build and boot on Pit - see that relocation still works correctly.

Change-Id: Ic9e55cb90b0f41c029d6a61081b0893d3a3e739a
Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/64133