efs: Display available stack size before relocation
When relocation is disabled, the U-Boot stack extends from the top of the
BSS section to the bottom of the malloc() region. In memory-constrained
environments the stack might become quite small depending on code size and
malloc size. So add an debug() to print the stack size before relocation.
This can help with monitoring the available memory.

BUG=chrome-os-partner:21115
BRANCH=pit
TEST=manual
Boot RO U-Boot with early-firmware-selection on Pit with '#define DEBUG'
at the top of common/board_f.c and see the stack size looks correct:

Available stack size 0000bd60

Change-Id: I70ed0a64d655c60dbd4a35329219a41b7302c9ca
Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/64134