gen: main: Process FDT options even without CONFIG_BOOTDELAY
At present when this option is not defined, we do not process the FDT
options - for example we do not run the boot command specified by the FDT.
Move this code into the common function and make sure it is always called
when CONFIG_OF_CONTROL is defined.

BUG=chrome-os-partner:21115
BRANCH=none
TEST=manual
Boot RO image on Pit and see that it now runs the boot command.

Change-Id: Ic0ab141f41602df51b6053c88edb801770a2994c
Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/64137