split: cros: fdt: Add a function to decode a region within memory
Some features require that their location in memory be fixed, for example
to pass data to the kernel or a second-stage U-Boot. These locations are
best described as offsets from the base of memory.

Systems often have more than one memory bank group (such as SDRAM and
on-chip SRAM), so the offset needs to specify the correct base. The
information about memory types is also stored in the device tree.

Add a function which decodes a memory region given the memory bank name
and the region type. So the kernel region could be specified as an offset
of 32KB into the start of SDRAM like this:

  kernel-memory = "/memory";
  kernel-offset = <0x8000>;

A configuration table at offset 4KB SRAM could be specified like this:

  configtable-memory = "/iram";
  configtable-offset = <0x1000>;

Normally these properties will be in /config but the caller can provide a
different node if require.

An optional suffix allows variants to be added. This allows U-Boot to adjust
the selected region based on circumstances. For example here we can select
either SDRAM or IRAM for 'cdata' by passing NULL or ",efs" as the suffix:

  cdata-memory = "/memory";
  cdata-offset = <0x80000>;

  cdata-memory,efs = "/iram";
  cdata-offset,efs = <0x1000>;

(sadly I had to add in a cros/ change here to keep pit happy. It gets
reverted by a later commit, so this is only to keep bisectability)

BUG=chrome-os-partner:21115
BRANCH=none
TEST=manual
Build and boot EFS on Pit and see that we can decode kernel and crossystem
areas correctly, and the system boots to the kernel.

Change-Id: I21992dea4f433e5033df71ae3e646a45443d327a
Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/64141