gen: fdt: Enhance flashmap function to deal with region properties
Flash regions can optionally be compressed or hashed. Add the ability to
read this information from the flashmap.

BUG=chromium:244019
BUG=chrome-os-partner:21115
TEST=manual
Adjust the RW-A-BOOT region to be compressed in the flashmap and see that
U-Boot correctly decompresses it.
Use the 'cros_test fmap' command to see that the flashmap entry information
is read correctly.

Change-Id: I68df7faa0022e8cd309003a16c53b5058b5dd916
Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/64142