efs: exynos: spl: Allow load address and offset to be specified
At present U-Boot is loaded from a fixed location in the boot media to a
fixed memory address. Add new machine params to specify these options.
Add messages to print out what is happening also, to avoid confusion.

BUG=chrome-os-partner:21115
BRANCH=pit
TEST=manual
Build and boot on Pit with EFS. See that we can now load a U-Boot from
any area of the SPI flash, and can load it either to IRAM or SDRAM.

Change-Id: I58b28d620428772ea61f256a2e3a22267b843bfd
Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/64145