efs: exynos: spl: Correct fast SPI reading algorithm to use offset
This function does use the offset but writes it out as a 32-bit word. For
reasons that I have not dived into this does not work.

Adjust the code to write the command as separate bytes, then switch to
word mode for reading the SPI flash. Also update the ordering of the
register writes slightly to better match the datasheet.

BUG=chrome-os-partner:21115
BRANCH=pit
TEST=manual
Build and boot on Pit with EFS. See that we can now load a U-Boot from
any area of the SPI flash, and can load it either to IRAM or SDRAM.

Signed-off-by: Simon Glass <sjg@chromium.org>

Change-Id: I24c9f8b073b4af396e228b9266215b674eb51860
Reviewed-on: https://gerrit.chromium.org/gerrit/64146
Reviewed-by: Simon Glass <sjg@chromium.org>
Tested-by: Simon Glass <sjg@chromium.org>