efs: cros: exynos: Support running SPL from U-Boot
With early-firmware-selection we want to run a RW SPL from U-Boot. This
requires that we don't init the SDRAM and clocks twice. Add the ability to
detect that U-Boot was the caller (rather than a reset) and avoid repeating
the init.

It is probably desirable to avoid initing SDRAM in RO SPL, just in case. A
bug is filed to track that discussion and decision.

BUG=chrome-os-partner:21115
BUG=chrome-os-partner:21511
BRANCH=pit
TEST=manual

Change-Id: Iea3d2ce7186ac4c32874d1ab395e83a2483164d8
Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/64147