efs: exynos: Support a hash of the data loaded by SPL
Add the facility to put a hash at the end of SPL, pointed to by the third
word in the variable-sized SPL header. This can then be accessed by SPL
and used to verify the U-Boot that it loads.

The hash is written into the SPL image by an external tool.

BUG=chrome-os-partner:21115
BRANCH=pit
TEST=manual

Change-Id: Ifa59b9e071136c8f839528f230db92ba3439ac4b
Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/64148