cros: Add key-based vboot flag driver
Sandbox does not have GPIOs, but it is convenient to be able to map keys
to vboot flags. For example the 'Pause/break' key can be used as the power
button. Add a new driver for this.

BUG=chromium:244019
TEST=manual
Boot U-Boot sandbox into the recovery screen by holding Escape on start-up.
Press Pause/Break and see that it shuts down.

Change-Id: Id7a95fb9ec1e3e40ae92625ab8fcf15d3bbf1314
Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/64151
Reviewed-by: Che-Liang Chiou <clchiou@chromium.org>