exynos: Add the concept of a board subrevision; add fixup map
On exynos5420-peach-pit the board revision strappings are actually
made up of two parts: the revision and the subrevision (which is
really a set of flags).  Add generic support for this concept to
U-Boot, adjusting existing code to deal with that fact.

On exynos5420-peach-pit we also had a misstuffing of revisions and
ended up duplicating a revision twice.  Add the concept of a revision
map to handle this.  This doesn't give us a way to actually
differentiate the two boards with the same stuffings (we'll just
deprecate the older one), but lets us continue to get a human-readable
revision number out of stuffings going forward.

BRANCH=none
BUG=chrome-os-partner:21481
TEST=Boot on rev3 and see rev3.0
TEST=Boot on rev4 and see rev4.0
TEST=Boot on snow PVT and see rev 2.0 (was rev 1)
TEST=Boot on snow misstuffed rev6 see rev 64.0 (was rev 6)
TEST=Boot on snow MP1.1 and see rev 3.1 (was rev 0)

Change-Id: I3a3e385510368c4634aa08fd0cbd8cb57236ce76
Signed-off-by: Doug Anderson <dianders@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/63994