exynos: gpio: add a bigger delay when reading board strappings
I saw cases on pit rev 3 where reading the board strappings would fail
with the 5us delay that was in the code.  Increase it to 15 to make it
reliable.  I specifically found that we could not properly read the
"Z-state" pins.

BRANCH=none
BUG=chrome-os-partner:21481
TEST=On rev3.0 pit (00ZZ), see that pins are left in a state were the
last two pins have pulldowns and the first have no pull.  Look at md.l
0x13400008 1 and see "5415".  These pins are GPY7_1-4, so look and see
that bits 2-9 are %00 0001 01.

Change-Id: I5822724604043449ff967bb6bd25ca380af01869
Signed-off-by: Doug Anderson <dianders@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/64338
Reviewed-by: Vadim Bendebury <vbendeb@chromium.org>