gen: image: Add feature to use the board rev to choose the fit image
Since the kernel image can contain multiple FDT images, the cleanest
way to deal with multiple board revisions in the kernel is to simply
have a single FDT per revision.  Then U-Boot can pick the correct FDT
and the kernel doesn't need any special "if (revision == XYZ)" hacks.

In U-Boot, we only get to include a single device tree and we make due
with "if (revision == XYZ)" hacks to avoid the pain of needing to
build a different U-Boot image for each revision of hardware.  ...but
the problem is that we have no way to pick the proper kernel image.

This CL introduces a way for U-Boot to pick the proper kernel image by
including a '#' in the compatible string of the U-Boot FDT.  For
instance, if we have revision 10 and the compatible string:
  compatible = "google,pit-rev#", "google,pit", ...
U-Boot will try to find a kernel FDT that matches "google,pit-rev10"

At the moment there is no global "get_revision" function in U-Boot, so
enabling this option for a board requires that the board define a
board_get_revision() function in its arch-specific includes.

CQ-DEPEND=CL:64392
BRANCH=none
BUG=chrome-os-partner:21481
TEST=Add a rev4 device tree to kernel and see it picked up on a rev4
pit.

Change-Id: I84be9e4cda6de4949900ec6c1b00350a752415b1
Signed-off-by: Doug Anderson <dianders@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/64385
Commit-Queue: Vadim Bendebury <vbendeb@chromium.org>
Reviewed-by: Vadim Bendebury <vbendeb@chromium.org>
Tested-by: Vadim Bendebury <vbendeb@chromium.org>