exynos5420: scale cpu frequency to max
As we donot do PMIC init in SPL, it is unsafe to clock the cpu at higher
clock then the clock recommanded at the default PMIC setting. But at the
same time we want to boot U-Boot at max possible clock to boot kernel at
high frequency supported. In order to do this bump up the cpu clock in
U-Boot as soon as we are done with PMIC initialization and adequate
voltage is set for the higher frequency.

CQ-DEPEND=CL:64700
BRANCH=pit
BUG=chrome-os-partner:21430
TEST=build and booted on peach, can see the APLL is set to 1.8GHz at
     U-Boot prompt
     Peach # md 0x10010100 1
     10010100: a0e10300
TEST=Stop U-Boot in SPL and using T-32 can see APLL in SPL is set to
     900MHz
TEST=suspend_stress_test several 100 times and warm reset several times

Signed-off-by: Alim Akhtar <alim.akhtar@samsung.com>

Change-Id: Ifd765784f07aad14e7afc6dcd383c0d6379bb8d7
Signed-off-by: Doug Anderson <dianders@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/64647
Commit-Queue: Alim Akhtar <alim.akhtar@samsung.com>
Reviewed-by: Alim Akhtar <alim.akhtar@samsung.com>
Tested-by: Alim Akhtar <alim.akhtar@samsung.com>